requirements.txt
openai>=1.0.0
python-dotenv
fastapi
uvicorn
gspread
oauth2client
openai
python-dotenv
